perf(client): replace the software renderer with a GPU-only backend (#478)
## Summary

- Replace the Classic client's software whole-window renderer with one
mandatory GPU-only rendering backend.
- Preserve map authority, painter order, fog/remembered disclosure,
lighting semantics, UI behavior, screenshots, and lifecycle recovery
while removing runtime CPU frame composition and fallback.
- Qualify dense large-view rendering, retained GPU resource behavior,
sparse map state, startup failure, and bounded device-loss recovery
against the issue's closure gates.

## Implementation / behavior

- Establish retained GPU device, asset, map, lighting, UI, text, and
presentation lifetimes on the production client path.
- Convert scene construction to backend-neutral retained commands and
compact GPU lighting/state updates without steady-state frame readback
or CPU-composed uploads.
- Remove the software renderer, window-surface presentation,
renderer-mode selection, and fallback paths once the GPU conformance
matrix is satisfied.

## Validation

- Final automated, conformance, runtime, hardware-backend, and
performance evidence will be recorded here after the implementation
converges on the final commit.
- The draft remains subject to complete-diff review, repository
validation, and required GitHub checks.

## Limitations / follow-up

- Qualified hardware performance evidence is platform-specific and will
identify exact revision, dirty state, device, driver, backend, shader
cohort, timing windows, uploads, allocations, and retained bytes.
- Minimap source/update-policy optimization remains outside this
delivery; its completed display still migrates to the GPU UI path.

Closes #477

<!--
atrinik-delivery:body:2240ea130af00600bbb197677feacd19e55a17347f4eec317eafd52f551a1980:start
-->
## Delivery status

Exact candidate head: `c3ec2042899ce4457b523ad95dea366a0003b85a`.

### Implementation / behavior

- The client uses one mandatory GPU-only whole-window renderer with
retained GPU map, lighting, UI, text, and presentation resources.
Software composition, window-surface presentation, renderer selection,
and runtime fallback paths are removed.
- Generated SPIR-V, DXIL, and MSL binaries are not tracked in Git or Git
LFS. The build/package pipeline acquires checksum-pinned DXC and
SPIRV-Cross inputs, compiles one 18-member cohort from the authoritative
HLSL source, verifies `client/shaders/SHA256SUMS`, and embeds the
validated cohort at build time.
- MAP2 continuation decoding is transactional across budget yields:
tiles, movement anchors, target selection, lighting, effects, and other
side effects publish together only after a complete successful replay.
Hard map clears, reconnects, replay failures, and GPU recovery discard
incomplete state.
- The retained map now has explicit publication validity, complete
wide/tall-sprite overscan and traversal, and correct initial/split-map
behavior. These changes restore missing objects and map edges after
login, teleports, and movement.
- Smooth lighting now uses projected per-pixel row lookup for floor and
roof sprites, a fixed structural-lighting row where appropriate, and
sparse changed-run uploads. The local player is forced fully opaque
before its first animation, and hidden/minimized/unfocused time no
longer advances presentation fades.
- The reported Windows startup crash was fixed by retaining the exact
book popup while other overlays are stacked above it.
- The one-click review launcher starts the complete server and client in
one flat bundle. The server is detached without a console and bound to
exactly one IPv4 loopback endpoint on UDP 1731, avoiding duplicate-stack
ownership and shutdown leaks.

### Review package

- `atrinik-classic-issue-477-windows-one-click-c3ec204.zip` is a
complete 228,368,212-byte client/server review package with 4,986
payload files, one `run-review.bat`, no separate client/server
directories, and a manifest bound to the exact candidate head.
- The launcher preflights UDP port 1731, creates isolated local server
state and a fresh QUIC identity, starts the server, waits for readiness,
then starts and connects the client to `127.0.0.1:1731`.
- SHA-256:
`7815974c34e0bb6b3b76eba8ad2fa21d8ca350321b1061b9b8375a4873e000c5`.

### Validation

- Fresh independent complete-diff review at the exact candidate head
reports zero actionable findings after rechecking MAP2 atomicity,
retained map/minimap validity, lighting, sprite visibility, recovery,
shaders, packaging, and launcher containment.
- Exact-head GitHub run
[33297360585](https://github.com/atrinik/classic/actions/runs/33297360585)
passes dependency verification, shader production, core, client, server,
integrated graph, trusted GPU renderer coverage, Windows
cross-build/package, the exact user-facing `run-review.bat` native
Windows smoke, and the required `Classic validation` aggregate.
- Exact-head CodeQL partitions, PR policy, and `codecov/patch` pass. The
ordinary opt-in GPU hardware qualification job skips by design.
- Local final-tree validation passes the client build, MAP2 state
harness, focused GPU/map/visibility/queue tests, all 42 locally
applicable client tests with 13 hardware/screenshot fixtures explicitly
skipped, all 17 libatrinik tests, shader manifest/embed validation,
deterministic package composition, ZIP CRC, and all 45 package/workflow
launcher tests.
- Exact-head qualification dispatch
[33289660032](https://github.com/atrinik/classic/actions/runs/33289660032)
was explicitly cancelled by the user after its hosted jobs passed but
its six Vulkan, Direct3D 12, and Metal cohorts remained unclaimed.
Hardware qualification is deferred until eligible runners exist.

### Limitations / follow-up

- Native Vulkan, Direct3D 12, and Metal reference/minimum hardware
qualification, performance cohorts, and human approval/population of
`gpu-approved-goldens.json` are explicitly deferred. The qualification
workflow remains opt-in for a future exact-head run; this delivery does
not claim those hardware gates passed.
-
[atrinik/devcontainer#38](https://github.com/atrinik/devcontainer/issues/38)
tracks public/fork-safe shader-capable images.
[atrinik/atrinik#514](https://github.com/atrinik/atrinik/issues/514)
tracks workspace supply-chain inventory for DXC, SPIRV-Cross, and image
evidence.
[atrinik/github-settings#77](https://github.com/atrinik/github-settings/issues/77)
tracks future qualified GPU runner provisioning.
- The follow-up issues were created under explicit authorization. Issue
#477 remains open. PR #478 is open, ready, cleanly mergeable, and
unmerged; maintainer approval is still required. The earlier
draft-status text above is retained as immutable initial-body history
and is superseded by this delivery status.

<!--
atrinik-delivery:body:2240ea130af00600bbb197677feacd19e55a17347f4eec317eafd52f551a1980:end
-->

+++ client/src/client/sound.c
@@ -1565,7 +1565,8 @@ int sound_test_main(const char *fixture_root) {
     printf("SDL3_mixer decoder contract: %s\n", decoders);
     SOUND_TEST_CHECK(sound_test_decode_fixture(fixture, &decoded_bytes));
     SOUND_TEST_CHECK(decoded_bytes == 23040);
-    printf("decoded %zu nonzero PCM bytes from Opus content at legacy .mid path\n", decoded_bytes);
+    printf("decoded %" PRIu64 " nonzero PCM bytes from Opus content at legacy .mid path\n",
+           (uint64_t)decoded_bytes);
     sound_background_hook_register(sound_test_background_hook);
 
     /* Uncached/cached starts, same-track policy updates, and explicit stop. */
